mcp/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
httpx==0.27.0
pydantic>=2.7.2
websockets==12.0
python-dotenv==1.0.1
mcp==1.8.1  # Model Context Protocol SDK